add continuous run mode enum
Summary:
In fbcode continuous, arch and build modes are defaulted to different values based on the continuous run mode. Example as below:

`opt` and `opt_hourly` - default mode is `opt`.
`aarch64` - everything should run in aarch64 regardless of the configs.

We need to propagate this field into the verifiable matcher to generate the right default values.

Reviewed By: pcerl017

Differential Revision: D54378899

fbshipit-source-id: 295b022b21bc76c91b47e0d68b68fb07d183335a