Forbid unsafe code
Summary:
There are no places where it's necessary for us to use `unsafe` in Citadel right now, so forbid it.

There are a few instances of `fbinit::assume_init()`, but I think we may be able to replace them with [`expect_init`](https://www.internalfb.com/D19456513) instead, which panics when initFacebook has not been called.

We will need to ensure that initFacebook is called in RELATES, our test runner for the RELATES adaptor, and any other binary which may construct Verse predicates but does not use `#[fbinit::main]` (either because the main function is in Python or for some other reason), since now they will encounter a panic (previously they encountered silent undefined behavior, presumably?).

Reviewed By: Acesine

Differential Revision: D54344160

fbshipit-source-id: 0b0abbec632419c3ca3a322a054c10e2e8bff0b8