Track the reason why an immediate change was detected
Summary: Add an `ImpactReason` enum indicating the path we took in `immediate_target_changes` to mark a target as affected, then propagate that reason through rdeps in `recursive_target_changes`. Carry the reason all the way through to BTD's output so that we can inspect it in Skycastle. Also summarize the number of affected targets with each reason in the BTD_SUCCESS event.

Reviewed By: ndmitchell

Differential Revision: D54436861

fbshipit-source-id: 800607b2796cd94f4f6965c0555e143cb468c9a2