Detect affected targets that are depending on removed targets
Summary:
## Problem
Currently if a target / package is removed, BTD doesn't include the affected targets that depend on those removed targets. I think we should treat those targets as affected.

## Discussion
Letting BTD return this kind of affected targets seems to be more consistent: one of their dependency is changed (here, it's removed), and should be rebuilt. User will get a build failure signal, which also seems to be more consistent experience (alternatively, this can be shown as a dangling dependency from a separate signal).

(In this diff, we still recursively looking for affected targets. A seem-to-work optimization is to only add targets directly depends on removed target, but I'm not sure about buck2's guarantee on implicit dependencies behaviors, eg. if buck enforces on transitive )

Reviewed By: ndmitchell, aniketmathur

Differential Revision: D56986908

fbshipit-source-id: 3225ecce1f4ebd03e59693b7cef518a6539ce742