Modify the internal structure of changes
Summary: Before the cell paths and project relative paths had to be consistent, but we couldn't tell that from the type. Now change the type so its immediately evident, which means we can optimise filter_by_extension (potentially making it more correct, if the function passed in is not pure) and allows us to write a generic filter.

Reviewed By: aniketmathur

Differential Revision: D57542251

fbshipit-source-id: 2ce9cee92e8eaa09430ebc5285017adcb526ecae