Separate Runway WWW continuous
Summary:
Right now, we hack in this 'force_shadow' logic. But this isn't ideal, as args.shadow is referenced in tons of places. This leads to runway-specific bugs like the one described in the D58031578

Let's separate Runway WWW into its own continuous TD for now, so we can actually pass args.shadow. We're doing this by adding a new run mode and adding a chronos config for it. Runway WWW targets will only be allowed in this run mode

Reviewed By: aniketmathur

Differential Revision: D58360956

fbshipit-source-id: c0e2dd66c6376fc8ad2eb63373297d91e3fec7d4