Return special exit code for scheduling failures
Summary:
If any jobs fail to schedule, scheduler will exit with a predefined exit code (100), which will be picked up by orchestrator and used to propagate warning to a diff.

Changing supertd is a bit messy, but the alternative is to use ```std::process::exit(code)```, which is generally not recommended.
I'll refactor it later to have a unified way to set workflow status and remove the ```.report()``` hack.

Reviewed By: Acesine

Differential Revision: D62506830

fbshipit-source-id: efde69c90e9d8041e5227b230d8f996669f8a56b