Use custom error type for setting workflow status
Summary:
Following up on the discussion with Acesine, converting custom result type to the custom error type for when we want to set non-standard workflow status.

This is more idiomatic, and allows seamless use of TD parts (e.g. BTD, Verse, etc) as libraries - it is now exposed as just Rust-standard  ```Result<(), CustomError>```

Reviewed By: Acesine

Differential Revision: D63149801

fbshipit-source-id: ab110510eb4283ea1aa4e8896b9ff7a1cd7194ef