Setup knob to control test scheduling
Summary:
**Context**
As we prepare for the mac capacity crunch, we want to be able to control the amount of test scheduling we do overall on the orchestrator. Lets set that up here using a [Just Knob](https://www.internalfb.com/intern/justknobs/?name=www%2Fdevx%2Fios_devxx#allow_test_scheduling) that decides a pass rate for us.

Current pass rates are:
 {F1886911959}

Reviewed By: jiawei-lyu

Differential Revision: D63325449

fbshipit-source-id: 6d3efb37c27bc1070f353e04db532aa7a5d6a4fa