btd: Refine impact reason
Summary:
There is an implicit priority here. Add a comment and reorder to reflect that.

source inputs changes take priority because they're mostly closely related to the
thing we want to validate i.e. a code change.

hashes come next because we want to protect the integrity of the target graph,
but might end up overtriggering on e.g. bzl file changes.

The rest are ordered based on blast radius. `ci_srcs` bypasses regular selection,
package detection will accept targets based on directory proximity to source changes.

Rules tend to over-trigger the most because their impact is unknown during the
parsing phase when the target graph is constructed. To determine impact, we
need to perform analysis but that is an expensive operation at scale, so we
defer it to the execution phase when the cost can be amortized across multiple targets.

Differential Revision: D64131598

fbshipit-source-id: d5b6ffac437c67388883607311436ed33245bfe0