td_util: Avoid unnecessary lossy string conversion in get_repo_root
Summary:
`String::from_utf8_lossy` for a path is unsafe because you'll end up
with a broken path if it's not UTF-8.

On all platforms, there's no need to pay for the `String` allocation
from `.replace` when all you want to do is remove the trailing new line.

On Unix, there's no need to pay for the `String` conversion at all
because paths need not be UTF-8.

Reviewed By: ndmitchell

Differential Revision: D65615072

fbshipit-source-id: b4fa540584d0a20519d742e75fa04042ddf6fe29