generalize event callsite for qe check logging
Summary:
It's currently hardcoded to log that it's a verse qe check event, but we want to do qe checke elsewhere as well.

The logging macro intentionally requires you to pass a value and not a variable to be logged, so instead, we log the callsite as a field in the data as a workaround.

Reviewed By: rjbailey

Differential Revision: D66670156

fbshipit-source-id: 1ae7419cdc680f23402f84ad44ce867a98b7363e