Return lower-valued exit status codes
Summary:
These were added in D62755198 to let us easily set workflow status from supertd. Unfortunately, [exit code 101 is also used by the rust runtime in its default handler for an uncaught panic](https://rust-cli.github.io/book/in-depth/exit-code.html), so it's not a good choice for classifying as Skipped. Since we accidentally classified uncaught panics as a Skipped step, we allowed the culprit diff causing S478885 to land.

There is no standard for exit status codes and I can't find a reference for the set reserved for the system, but I know we at least need to avoid 101 (uncaught rust panic) and 129-159 (terminated by signal). This diff changes supertd to emit the codes expected by the classifiers added in the previous diff.

Differential Revision: D67777531

fbshipit-source-id: 65cd98f9054b6f45d571c7e1540cada8078f053a