add verse run for guaranteed counts per PG
Summary: In order to get the guaranteed verifiable selection ratio, we need to count the number of verifiables per PG that would be run under guaranteed thresholds. This means evaluating verse again with the additional runtime context flag "apply_dr_control_min_value_override" so that we use the min threshold in the cumulative cap predicate.

Reviewed By: Acesine

Differential Revision: D70366520

fbshipit-source-id: 10dde5ec100fc4ffbe9392169e5884594ba2641a