Extract step fn to make logging easier across scheduler
Summary:
Currently every module has its own definition of `step()` as a closure, binding to an `Instant` set to the start timestamp. If some function needs to use `step`, it has to be passed down as `impl Fn()`, which is unpleasant.
This diff extracts `step()` as a utility fn, accessible to all modules, and defines `START_TIME` explicitly, initializing on binary start.

We lose a bit of flexibility because now there is no way to set START_TIME manually, but I think it's a good thing, because places that used to set it later, like ranker or relates, will now have the timestamps in log consistent with e.g. btd or scheduler.

Reviewed By: aniketmathur

Differential Revision: D70560078

fbshipit-source-id: 61c0077ab97a30b59115b7476c0b6ff35067f07c