make read_file_lines_parallel more efficient
Summary:
In D70700296, I didn't realize this was used in TD (bad grepping), so I
introduced a somewhat lazy fix just sorting things after processing them.

The more efficient way to do this is to just fill our buffer out of order,
which is what this does. It does, however, require a bit of unsafe, since we
need to be sure that we've filled the whole buffer.

I also tried to do this with a channel to reduce the time spent with a mutex
held, but I didn't see a meaningful improvement here (in fact it seemed worse),
so mutex it is

Reviewed By: aniketmathur

Differential Revision: D71984800

fbshipit-source-id: 9e60376148edf6041375165059866b0673d04cc6