Skip recursion
Summary:
# What

For continuous runs, where we don't pass in a base graph, everything is marked as affected. Since everything is already affected, we don't need to traverse the graph and look at all their dependencies.

67% faster runtime
6% less memory

# Changes

This does alter the output, in that terminal nodes are no longer tagged. This is okay since for continuous runs we select everything anyway, and the intuition behind terminal nodes makes less sense when everything is new.

Reviewed By: aniketmathur

Differential Revision: D72462579

fbshipit-source-id: 7bb60b418e52a71d9cd2a69109d43a66fd3063fe