Log reasons why signals ran
Summary:
For [why signals ran](https://fb.workplace.com/groups/1243087773044302/posts/1629798141039928), we need to log the selection reasons. We want to do this after verse so that we only log the targets that are actually going to run

Im doing this at the start of the scheduler b/c we cant get a mapping (of jobs:targets) after the targets have been packed into jobs. This will technically over-log since we can have failures during scheduling, but it seems too much overhead to keep a mapping of the jobs:targets and then figure out which targets to run after their job has been scheduled

Start a separate thread to perform the logging so we dont block the main process

Reviewed By: aniketmathur

Differential Revision: D74094960

fbshipit-source-id: 42a4a2c34cf69f977abf54a44188eff8f7af77c1