Adding original labels to BTD output for easier proccessing for change in labels
Summary: I want to change the output of BTD to also include information of the original labels of a target, this is to help later on when outputting the information of changed labels later on. This will only add old label information if the reason for the target is labels. New fields `added_labels` and `removed_labels` in reason.

Differential Revision: D76533096

fbshipit-source-id: 42da91bd1b43dc1725ba11acdd95994d0833e985