Change 'labels' reason to only show up when there is a change ci labels specifically
Summary: 'labels' reason shows up when there is a change in labels. changing this so that this reason only shows up when ci labels specifically are changed. This will help in future tasks related to the user empowrment project. brought back package_value reason to represent a package value change (which will be a label change that is not a ci label change)

Differential Revision: D76458700

fbshipit-source-id: 37dfdf6528ebe6a2231d360a18a3e420131e7f42