set `OMP_NUM_THREADS=1`
Summary:
Like it says in the title. We don't really need parallelism here (it's mostly
useful for speeding up predictions), and if we don't set this then we end up
spawning way too many threads (18K total on my machine!)

I tried initially to do this by not enabling OpenMP at compile time, but I was
not able to get that to work and got linker errors. It's not clear if that's
even supported. This also had the downside of forking the graph, which felt
not great.

Reviewed By: StanislawSwierc

Differential Revision: D77221368

fbshipit-source-id: 6a7091f1c262ec09a26256bfd5dc7bd011ff7b23