Add Dangling Check to BTD
Summary:
**Stack**
This stack introduces support for dangling check in the BTD step of the workflow. The check is limited to the universe that is provided for the given project.

This change aims to solve two problems that existed with the previous dangling check implementation:
1. Dangling Check Failure resulted in the failure of the workflow and no further steps were computed.
2. Dangling check used to happen across all the universe in the flagfile for the given project. This created a lot of false positives for the xplat project.

The change solves the problem 1 by writing the dangling errors to a json file and processing(as suggested by aniketmathur) it in a separate step to send the signal to SLP as a warning. Problem 2 is solved by explicitly providing the universe(as suggested by ndmitchell) for which we want the dangling check to happen for the given project.

Currently, fbandroid projects are doing explicit dangling check outside of citadel using [validate-fbandroid-buck-graph-uquery](https://www.internalfb.com/sandcastle/workflow/2427440199163651611) which is not the ideal way and it should happen in the citadel as detailed in this [post](https://fb.workplace.com/groups/targetdeterminator/permalink/4824415247784927/) and comments. After rolling this feature out we can work with `android_repo_parseability` and see if we can  deprecate the validate-fbandroid-buck-graph-uquery

**Diff**

Introduces new args `check_dangling_universe` and `write_dangling_errors_to_file` in BTD.  `check_dangling_universe` is meant for the specific universe in which we want the dangling check and `write_dangling_errors_to_file` is the dangling error file. This file will be used for further processing to send dangling error to SLP.

Reviewed By: krallin

Differential Revision: D84888907

fbshipit-source-id: 317831a83f008feba59e7b94d9e12871da4dae90