Update target.rs to use tests instead of buck.tests
Summary:
**Stack**

This stack is to enable deserializing the buck's tests attribute which can either be an array or a select statement. This deserialized field will be used to catch the dangling dependencies in the tests. Currently, the dangling check is not working for tests because buck.tests field doesn't exist as a result dangling failure are not being caught for tests.

For fields(such as label) that can have select statements in addition to simple array, we already do the deserialization to flatten. We are refactoring the code so that it can be re-used for tests as well.

Finally this will enable us to catch dangling errors in tests.

**Diff**

Update the target.rs to use tests attribute instead of non existent buck.tests and using the custom serializer to get list of targets even when tests are a select statement

Differential Revision: D88872529

fbshipit-source-id: 47827765e2f1dcee9b38c79dba9a6b9e9c647a97