flattened_select.rs which can be used to flatten the select statements in tests
Summary:
**Stack**
This stack is to enable deserializing the buck's tests attribute which can either be an array or a select statement. This deserialized field will be used to catch the dangling dependencies in the tests. Currently, the dangling check is not working for tests because buck.tests field doesn't exist as a result dangling failure are not being caught for tests.

For fields(such as label) that can have select statements in addition to simple array, we already do the deserialization to flatten. We are refactoring the code so that it can be re-used for tests as well.

Finally this will enable us to catch dangling errors in tests.
**Diff**
create flattened_select.rs which can be used to flatten the select statements in tests.

Differential Revision: D89190808

fbshipit-source-id: 18ea20f51dfdbaa4e972bc993bf54a93d0601167