Introduce deserialization to get list out of select statements and refactor label.rs
Summary:
**Stack**

This stack is to enable deserializing the buck's tests attribute which can either be an array or a select statement. This deserialized field will be used to catch the dangling dependencies in the tests. Currently, the dangling check is not working for tests because buck.tests field doesn't exist as a result dangling failure are not being caught for tests.

For fields(such as label) that can have select statements in addition to simple array, we already do the deserialization to flatten. We are refactoring the code so that it can be re-used for tests as well.

Finally this will enable us to catch dangling errors in tests.

**Diff**

Create a select.rs that uses Visitor Pattern as in label.rs and expose the methods publicly to be re-used.
This is select.rs is then used in label.rs

Differential Revision: D88872476

fbshipit-source-id: 3d11d0f9d807a542bac0479edfef2a88c4396b6f